chore: moving is_ldoi from state_props to matrix_props
