+++ toqito/matrix_props/__init__.py
@@ -39,3 +39,4 @@ from toqito.matrix_props.commutant import commutant
 from toqito.matrix_props.mutual_coherence import mutual_coherence
 from toqito.matrix_props.is_absolutely_k_incoherent import is_absolutely_k_incoherent
 from toqito.matrix_props.is_k_incoherent import is_k_incoherent
+from toqito.matrix_props.is_ldoi import is_ldoi
