+++ toqito/state_props/__init__.py
@@ -26,7 +26,6 @@ from toqito.state_props.is_antidistinguishable import is_antidistinguishable
 from toqito.state_props.is_distinguishable import is_distinguishable
 from toqito.state_props.is_unextendible_product_basis import is_unextendible_product_basis
 from toqito.state_props.is_sic_povm import is_sic_povm
-from toqito.state_props.is_ldoi import is_ldoi
 from toqito.state_props.common_quantum_overlap import common_quantum_overlap
 from toqito.state_props.renyi_entropy import renyi_entropy
 from toqito.state_props.learnability import learnability
